Adjust ID token handling for proxy mode
In proxy mode, the `id_token` from the upstream identity provider is
now mapped to `access_token` in the response. This ensures that the
token is verifiable, whereas the upstream access token might be from a
different issuer.

Without this fix, authentication to MCP-Trino appears to work, but upon
making any requests, there are token verification failures.